requirements.txt
# Core dependencies
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
rich>=10.0.0
PyYAML>=5.4.0
aiohttp>=3.8.0

# AI and ML
scikit-learn==1.0.2
numpy==1.21.6
pandas==1.3.5
torch>=2.0.0
transformers>=4.15.0
sentence-transformers>=2.2.0

# Testing
pytest>=6.2.5
pytest-cov>=2.12.1
pytest-asyncio>=0.16.0
pytest-mock>=3.6.1
pytest-xdist>=2.4.0
pytest-timeout>=2.1.0
pytest-env>=0.8.0
pytest-randomly>=3.8.0
pytest-sugar>=0.9.4
pytest-html>=3.1.1
pytest-metadata>=1.11.0
pytest-clarity>=1.0.1
pytest-benchmark>=3.4.1
pytest-profiling>=1.7.0
pytest-xprocess>=0.18.0
pytest-rerunfailures>=10.2
pytest-repeat>=0.9.1

# Git integration
gitpython>=3.1.24
pygit2>=1.7.0

# Browser testing
selenium==4.0.0
playwright==1.16.1

# Monitoring
prometheus-client>=0.11.0
python-json-logger>=2.0.2
structlog>=21.1.0
loguru>=0.5.3
sentry-sdk>=1.3.1
opentelemetry-api>=1.7.1
opentelemetry-sdk>=1.7.1
opentelemetry-instrumentation-fastapi>=0.27b0
opentelemetry-exporter-otlp>=1.7.1
opentelemetry-exporter-jaeger>=1.7.1
opentelemetry-exporter-zipkin>=1.7.1
opentelemetry-exporter-prometheus>=0.54b0
opentelemetry-exporter-otlp-proto-grpc>=1.7.1
opentelemetry-exporter-otlp-proto-http>=1.7.1
opentelemetry-instrumentation>=0.27b0
opentelemetry-instrumentation-requests>=0.27b0
opentelemetry-instrumentation-urllib3>=0.27b0
opentelemetry-instrumentation-urllib>=0.27b0

# Documentation
sphinx>=4.2.0
sphinx-rtd-theme>=1.0.0
sphinx-autodoc-typehints>=1.12.0
sphinx-copybutton>=0.4.0
sphinx-tabs>=3.2.0
sphinx-panels>=0.6.0
sphinx-issues>=1.2.0
sphinx-github-changelog>=1.0.0
sphinx-autobuild>=2021.3.14

# Security
bandit>=1.7.0
safety>=1.10.3
cryptography>=35.0.0
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=3.2.0
python-multipart>=0.0.5
itsdangerous>=2.0.1
pyjwt>=2.1.0
oauthlib>=3.1.1
requests-oauthlib>=1.3.0
authlib>=0.15.5

# Code quality
black>=21.9b0
isort>=5.9.3
flake8>=3.9.2
mypy>=0.910
pylint>=2.9.6
radon>=5.1.0
xenon>=0.7.3
vulture>=2.3
pytype>=2021.9.9
pyright>=1.1.194
pyre-check>=0.9.12
pydocstyle>=6.1.1
docformatter>=1.4.0
autopep8>=1.5.7
yapf>=0.31.0
prospector>=1.5.2
pylama>=8.3.7
wemake-python-styleguide>=0.15.3
flake8-docstrings>=1.6.0
flake8-bugbear>=21.9.2
flake8-comprehensions>=3.7.0
flake8-simplify>=0.14.3
flake8-print>=4.0.0
flake8-import-order>=0.18.1
flake8-quotes>=3.3.0
flake8-multiline-containers>=0.0.18
flake8-pytest-style>=1.6.0

# Development
ipython==7.28.0
ipdb==0.13.9
watchdog==2.1.6

# Async support
aiofiles>=0.8.0
asyncio==3.4.3
websockets>=10.0

# Formatting and parsing
Markdown==3.5
Pygments==2.10.0
beautifulsoup4==4.12.2

# Database
sqlalchemy>=1.4.23
alembic>=1.7.3
psycopg2-binary>=2.9.1
asyncpg>=0.24.0
aiosqlite>=0.17.0
motor>=2.5.1
redis>=3.5.3
aioredis>=2.0.0
pymongo>=3.12.0
elasticsearch>=7.15.0
elasticsearch-async>=6.2.0
cassandra-driver>=3.25.0
aiocassandra>=0.7.0
neo4j>=4.3.0
neo4j-driver>=4.3.0
aiomysql>=0.0.21
aiopg>=1.2.2

# API and Web
requests>=2.26.0
httpx>=0.19.0
graphql-core>=3.2.0
strawberry-graphql>=0.96.0
graphene>=3.0.0
ariadne>=0.14.0
fastapi-utils>=0.2.1
fastapi-cache2>=0.1.9
fastapi-limiter>=0.1.4
fastapi-pagination>=0.12.0
fastapi-sqlalchemy>=0.2.1
fastapi-users>=9.3.2
fastapi-mail>=1.0.1

# Utilities
python-dotenv>=0.19.0
pytz>=2021.1
python-dateutil>=2.8.2

# Documentation
mkdocs==1.2.3
mkdocs-material==7.3.6

# Monitoring
grafana-api-client==0.2.0
prometheus-api-client==0.5.1

# GitHub integration
PyGithub==1.59.1
